remove utils from conceptmap
